ipfs-cluster-0.12.0-alpha1
```
[16:03:58] ~/i/distributions (cluster-0.12.0-alpha1|✚3…) $ ipfs object diff /ipns/dist.ipfs.io (tail -n1 versions)
~ Qmbm2UJyhAXTKv65hLwHYrBfxyVb7MENQaADB6utmui1BY QmZPUy8b9iULke4JJ4wo6UdUP2zQNkjxFrRnv8RszzViry "index.html"
~ QmctZcVV1G78AC9tGNnrDF3g23c5559zdmiy6s34pvUckT QmQR7gNEKWcM2qA4X7yChftJcmGywVABYJWHNUcUv99Bc9 "ipfs-cluster-ctl/versions"
+ QmZpgnm3zuuaWZZJpJpDUDPQx8cTfVKQV79hj7UPicPfGq "ipfs-cluster-ctl/v0.12.0-alpha1"
~ QmPQxrxTYj7gxbea6H35pcAJttWLRkt56gaQK3YMKPPMaF QmNg59iqv1MMB8h3GwgRWhSJ55mcQPgagQa3yNBMU7sCVC "ipfs-cluster-service/versions"
+ QmZjD7m257CwnDVQU9B1zN2PhtoyCo4o4Hk5fyAkkv6EUk "ipfs-cluster-service/v0.12.0-alpha1"
~ QmTS4sMh52nvBU2qzouAiKzCa8phijsdYwv3uBW3judheq QmPgnfFXi1DqWX1DuSCCPGempbvWAt8YPq6KL4Rufmmf7G "sitemap.xml"
+ QmZkahY85vzJg9FuW74YcTMRpwFuVNrtngHqoBQYoQVBQQ "ipfs-cluster-follow"
```

Note change to build-go.sh: it tries to `cd` to the $package in order to clean the repo.
Things work as long as that folder exists in master but not when it does not exist because
it has been added on a non-merged tag. In the end, it does not actually need to
`cd` into that subfolder, as it can do the cleaning and branch checkout from the top-level
and then buildWithMatrix receives the package name and does things normally (I think).