Fix case where 'ipfs object diff' returns a dir
I had assumed that 'ipfs object diff' only returned files but it can
return directories too, in which case 'ipfs cat' fails, so this
switches (back) to 'ipfs get' to recursively fetch the contents of the
changed directories.

Also throw an exception when the 'diff' command exits with
code=2 (indicates some unexpected error).