chore: remove nightly status from commits
TLDR: Nightlies create issue for every build failure, so we do not need this
anymore.

Longer rationale:

This made sense only when we were testing nightlies on a PR.
Now that they run once a day, we should not attach build status to
commits, because there is no way to remove "Problem with" message.
We could make it green, but "Problem with" remains confusing.