Do not include full path to binary in .sha512 files on darwin
See https://github.com/ipfs/kubo/issues/9323.

This is a darwin release thing because the signature process for macOS means
that we recalculate sha512 and cid and we do this using an absolute path so
the result includes the path to the file.

The fix is to cd to the folder and do it from there.